refactoring; remove obsolete files
